[CCUBE-2183][GZ] rename form-histogram, fix lint err
