+++ src/form/form-histogram-slider/form-histogram-slider.tsx
