+++ src/form/form-histogram-slider/index.ts
@@ -1,2 +1,2 @@
-export { FormHistogramSlider } from "./form-histogram";
+export { FormHistogramSlider } from "./form-histogram-slider";
 export type { FormHistogramSliderProps } from "./types";
